Temporary dependensy removal
Removed the Sleuth and Zipkin dependecies because they caused a TimeoutException in the "testGetFoodNutrients" unit test.
It seems that they caused an intrinsic Hystric resiliency failure, because the call to USDA took almost 7.5s (more that the 5s default of Hystrix).

The dependecies will be added together with the Hystric parameter tuning.

Signed-off-by: Cristina Dinescu <crina91@gmail.com>